+++ examples/Components/Subnavigation/index.vue
@@ -51,12 +51,6 @@
     <router-link class="subnavigation__link" to="/collaboration">
       Collaboration
     </router-link>
-    <router-link class="subnavigation__link" to="/collaboration2">
-      Collaboration 2
-    </router-link>
-    <router-link class="subnavigation__link" to="/collaboration3">
-      Collaboration 3
-    </router-link>
   </div>
 </template>
